Split validation engine internals
