+++ packages/validation/src/validation.ts
@@ -1,34 +1,14 @@
-import validator from 'validator';
-
 import type {
   Constructor,
   MetadataPropertyKey,
 } from '@fluojs/core';
-import {
-  getClassValidationRules,
-  getDtoBindingSchema,
-  getDtoValidationSchema,
-  type ClassValidationRule,
-  type DtoFieldBindingMetadata,
-  type DtoFieldValidationRule,
-  type ValidationIssueMetadata,
-  type ValidationRuleResult,
-} from '@fluojs/core/internal';
+import type { ClassValidationRule, DtoFieldBindingMetadata, DtoFieldValidationRule, ValidationIssueMetadata, ValidationRuleResult } from '@fluojs/core/internal';
 
 import { DtoValidationError } from './errors.js';
+import { getCachedDtoMetadata, resolveNestedDto } from './internal/dto-metadata-cache.js';
+import { assignSafeOwnEnumerableProperties, getIterableValues, isPlainObject } from './internal/object-utils.js';
+import { getRuleHandler, type NonCustomRule } from './internal/rule-handlers.js';
 import type { ValidationIssue, Validator } from './types.js';
-
-function isClassConstructor(dto: Constructor | (() => Constructor)): dto is Constructor {
-  return typeof dto === 'function' && Function.prototype.toString.call(dto).startsWith('class ');
-}
-
-function resolveNestedDto(dto: Constructor | (() => Constructor)): Constructor {
-  if (isClassConstructor(dto)) {
-    return dto;
-  }
-
-  return (dto as () => Constructor)();
-}
 
 function toFieldName(propertyKey: MetadataPropertyKey): string {
   return typeof propertyKey === 'string' ? propertyKey : String(propertyKey);
@@ -68,45 +48,6 @@ function normalizeResult(
   return [normalizeIssue(result as ValidationIssueMetadata, field, source)];
 }
 
-function getIterableValues(value: unknown): unknown[] | undefined {
-  if (Array.isArray(value)) return value;
-  if (value instanceof Set) return Array.from(value.values());
-  if (value instanceof Map) return Array.from(value.values());
-  return undefined;
-}
-
-function isPlainObject(value: unknown): value is Record<PropertyKey, unknown> {
-  if (typeof value !== 'object' || value === null || Array.isArray(value)) {
-    return false;
-  }
-
-  const prototype = Object.getPrototypeOf(value);
-  return prototype === Object.prototype || prototype === null;
-}
-
-const DANGEROUS_KEYS = new Set(['__proto__', 'constructor', 'prototype']);
-
-function assignSafeOwnEnumerableProperties(
-  target: Record<PropertyKey, unknown>,
-  source: Record<PropertyKey, unknown>,
-): void {
-  for (const key of Reflect.ownKeys(source)) {
-    if (typeof key === 'string' && DANGEROUS_KEYS.has(key)) {
-      continue;
-    }
-
-    if (!Object.prototype.propertyIsEnumerable.call(source, key)) {
-      continue;
-    }
-
-    target[key] = source[key as keyof typeof source];
-  }
-}
-
-function isEmptyValue(value: unknown): boolean {
-  return value === '' || value === null || value === undefined;
-}
-
 function joinFieldPath(parent: string, child?: string): string {
   if (!child) return parent;
   return child.startsWith('[') ? `${parent}${child}` : `${parent}.${child}`;
@@ -118,22 +59,6 @@ function prefixIssues(
   source: ValidationIssue['source'],
 ): ValidationIssue[] {
   return issues.map((issue) => ({ ...issue, field: joinFieldPath(fieldPrefix, issue.field), source: issue.source ?? source }));
-}
-
-type RuleKind = DtoFieldValidationRule['kind'];
-type NonCustomRule = Exclude<DtoFieldValidationRule, { kind: 'custom' | 'nested' }>;
-type DtoValidationSchema = ReturnType<typeof getDtoValidationSchema>;
-
-interface CachedDtoMetadata {
-  bindingMap: Map<MetadataPropertyKey, DtoFieldBindingMetadata>;
-  classValidationRules: ReturnType<typeof getClassValidationRules>;
-  dtoValidationSchema: DtoValidationSchema;
-  mergedPropertyKeys: Set<MetadataPropertyKey>;
-  nestedDtoTransforms: readonly {
-    each: boolean;
-    propertyKey: MetadataPropertyKey;
-    target: Constructor;
-  }[];
 }
 
 interface NestedTraversalContext {
@@ -165,269 +90,6 @@ function exitTraversal(value: unknown, context?: NestedTraversalContext): void {
 
   context.active.delete(value);
 }
-
-const dtoMetadataCache = new WeakMap<Constructor, CachedDtoMetadata>();
-
-function collectNestedDtoTransforms(dtoValidationSchema: DtoValidationSchema): CachedDtoMetadata['nestedDtoTransforms'] {
-  const nestedEntries: CachedDtoMetadata['nestedDtoTransforms'][number][] = [];
-
-  for (const entry of dtoValidationSchema) {
-    const nestedRule = entry.rules.find(
-      (rule: DtoFieldValidationRule): rule is Extract<DtoFieldValidationRule, { kind: 'nested' }> => rule.kind === 'nested',
-    );
-
-    if (!nestedRule) {
-      continue;
-    }
-
-    nestedEntries.push({
-      each: nestedRule.each === true,
-      propertyKey: entry.propertyKey,
-      target: resolveNestedDto(nestedRule.dto),
-    });
-  }
-
-  return nestedEntries;
-}
-
-function getCachedDtoMetadata(target: Constructor): CachedDtoMetadata {
-  const cached = dtoMetadataCache.get(target);
-
-  if (cached) {
-    return cached;
-  }
-
-  const bindingMap = getDtoBindingMap(target);
-  const dtoValidationSchema = getDtoValidationSchema(target);
-  const classValidationRules = getClassValidationRules(target);
-  const mergedPropertyKeys = new Set<MetadataPropertyKey>([
-    ...bindingMap.keys(),
-    ...dtoValidationSchema.map((entry: { propertyKey: MetadataPropertyKey }) => entry.propertyKey),
-  ]);
-  const nestedDtoTransforms = collectNestedDtoTransforms(dtoValidationSchema);
-  const next: CachedDtoMetadata = {
-    bindingMap,
-    classValidationRules,
-    dtoValidationSchema,
-    mergedPropertyKeys,
-    nestedDtoTransforms,
-  };
-
-  dtoMetadataCache.set(target, next);
-  return next;
-}
-
-type RuleHandler<K extends RuleKind> = {
-  defaultCode: string;
-  describe: (field: string, rule: Extract<DtoFieldValidationRule, { kind: K }>) => string;
-  validate: (rule: Extract<DtoFieldValidationRule, { kind: K }>, value: unknown) => boolean;
-};
-
-function getRuleHandler<K extends RuleKind>(rule: Extract<DtoFieldValidationRule, { kind: K }>): RuleHandler<K> {
-  return RULE_HANDLERS[rule.kind] as RuleHandler<K>;
-}
-
-const RULE_HANDLERS: { [K in RuleKind]: RuleHandler<K> } = {
-  validateIf: {
-    defaultCode: 'VALIDATE_IF',
-    describe: (field) => `${field} is conditionally invalid.`,
-    validate: () => true,
-  },
-  defined: {
-    defaultCode: 'REQUIRED',
-    describe: (field) => `${field} is required.`,
-    validate: (_rule, value) => value !== undefined && value !== null,
-  },
-  optional: {
-    defaultCode: 'OPTIONAL',
-    describe: (field) => `${field} is optional.`,
-    validate: () => true,
-  },
-  equals: {
-    defaultCode: 'EQUALS',
-    describe: (field, rule) => `${field} must equal ${String(rule.value)}.`,
-    validate: (rule, value) => value === rule.value,
-  },
-  notEquals: {
-    defaultCode: 'NOT_EQUALS',
-    describe: (field, rule) => `${field} must not equal ${String(rule.value)}.`,
-    validate: (rule, value) => value !== rule.value,
-  },
-  empty: {
-    defaultCode: 'EMPTY',
-    describe: (field) => `${field} must be empty.`,
-    validate: (_rule, value) => isEmptyValue(value),
-  },
-  notEmpty: {
-    defaultCode: 'NOT_EMPTY',
-    describe: (field) => `${field} should not be empty.`,
-    validate: (_rule, value) => !isEmptyValue(value),
-  },
-  in: {
-    defaultCode: 'IN',
-    describe: (field) => `${field} must be one of the allowed values.`,
-    validate: (rule, value) => rule.values.includes(value),
-  },
-  notIn: {
-    defaultCode: 'NOT_IN',
-    describe: (field) => `${field} contains a forbidden value.`,
-    validate: (rule, value) => !rule.values.includes(value),
-  },
-  string: {
-    defaultCode: 'INVALID_STRING',
-    describe: (field) => `${field} must be a string.`,
-    validate: (_rule, value) => typeof value === 'string',
-  },
-  number: {
-    defaultCode: 'INVALID_NUMBER',
-    describe: (field) => `${field} must be a number.`,
-    validate: (rule, value) => typeof value === 'number' && (rule.allowNaN || !Number.isNaN(value)),
-  },
-  boolean: {
-    defaultCode: 'INVALID_BOOLEAN',
-    describe: (field) => `${field} must be a boolean.`,
-    validate: (_rule, value) => typeof value === 'boolean',
-  },
-  date: {
-    defaultCode: 'INVALID_DATE',
-    describe: (field) => `${field} must be a Date instance.`,
-    validate: (_rule, value) => value instanceof Date && !Number.isNaN(value.getTime()),
-  },
-  array: {
-    defaultCode: 'INVALID_ARRAY',
-    describe: (field) => `${field} must be an array.`,
-    validate: (_rule, value) => Array.isArray(value),
-  },
-  object: {
-    defaultCode: 'INVALID_OBJECT',
-    describe: (field) => `${field} must be an object.`,
-    validate: (_rule, value) => isPlainObject(value),
-  },
-  enum: {
-    defaultCode: 'INVALID_ENUM',
-    describe: (field) => `${field} must be a supported enum value.`,
-    validate: (rule, value) => rule.values.includes(value),
-  },
-  int: {
-    defaultCode: 'INVALID_INT',
-    describe: (field) => `${field} must be an integer.`,
-    validate: (_rule, value) => typeof value === 'number' && Number.isInteger(value),
-  },
-  divisibleBy: {
-    defaultCode: 'DIVISIBLE_BY',
-    describe: (field, rule) => `${field} must be divisible by ${String(rule.value)}.`,
-    validate: (rule, value) => typeof value === 'number' && !Number.isNaN(value) && value % rule.value === 0,
-  },
-  positive: {
-    defaultCode: 'POSITIVE',
-    describe: (field) => `${field} must be positive.`,
-    validate: (_rule, value) => typeof value === 'number' && value > 0,
-  },
-  negative: {
-    defaultCode: 'NEGATIVE',
-    describe: (field) => `${field} must be negative.`,
-    validate: (_rule, value) => typeof value === 'number' && value < 0,
-  },
-  min: {
-    defaultCode: 'MIN',
-    describe: (field, rule) => `${field} must be greater than or equal to ${String(rule.value)}.`,
-    validate: (rule, value) => typeof value === 'number' && !Number.isNaN(value) && value >= rule.value,
-  },
-  max: {
-    defaultCode: 'MAX',
-    describe: (field, rule) => `${field} must be less than or equal to ${String(rule.value)}.`,
-    validate: (rule, value) => typeof value === 'number' && !Number.isNaN(value) && value <= rule.value,
-  },
-  minDate: {
-    defaultCode: 'MIN_DATE',
-    describe: (field, rule) => `${field} must be on or after ${rule.value.toISOString()}.`,
-    validate: (rule, value) => value instanceof Date && !Number.isNaN(value.getTime()) && value.getTime() >= rule.value.getTime(),
-  },
-  maxDate: {
-    defaultCode: 'MAX_DATE',
-    describe: (field, rule) => `${field} must be on or before ${rule.value.toISOString()}.`,
-    validate: (rule, value) => value instanceof Date && !Number.isNaN(value.getTime()) && value.getTime() <= rule.value.getTime(),
-  },
-  contains: {
-    defaultCode: 'CONTAINS',
-    describe: (field, rule) => `${field} must contain ${rule.value}.`,
-    validate: (rule, value) => typeof value === 'string' && value.includes(rule.value),
-  },
-  notContains: {
-    defaultCode: 'NOT_CONTAINS',
-    describe: (field, rule) => `${field} must not contain ${rule.value}.`,
-    validate: (rule, value) => typeof value === 'string' && !value.includes(rule.value),
-  },
-  length: {
-    defaultCode: 'LENGTH',
-    describe: (field) => `${field} must have a valid length.`,
-    validate: (rule, value) => typeof value === 'string' && value.length >= rule.min && (rule.max === undefined || value.length <= rule.max),
-  },
-  minLength: {
-    defaultCode: 'MIN_LENGTH',
-    describe: (field, rule) => `${field} must have length at least ${String(rule.value)}.`,
-    validate: (rule, value) => typeof value === 'string' && value.length >= rule.value,
-  },
-  maxLength: {
-    defaultCode: 'MAX_LENGTH',
-    describe: (field, rule) => `${field} must have length at most ${String(rule.value)}.`,
-    validate: (rule, value) => typeof value === 'string' && value.length <= rule.value,
-  },
-  nested: {
-    defaultCode: 'INVALID_NESTED',
-    describe: (field) => `${field} contains invalid nested data.`,
-    validate: () => true,
-  },
-  validatorjs: {
-    defaultCode: 'INVALID_FIELD',
-    describe: (field) => `${field} is invalid.`,
-    validate: (rule, value) => runValidatorJs(rule, value),
-  },
-  arrayContains: {
-    defaultCode: 'ARRAY_CONTAINS',
-    describe: (field) => `${field} must contain the required values.`,
-    validate: (rule, value) => Array.isArray(value) && rule.values.every((expected: unknown) => value.includes(expected)),
-  },
-  arrayNotContains: {
-    defaultCode: 'ARRAY_NOT_CONTAINS',
-    describe: (field) => `${field} contains forbidden values.`,
-    validate: (rule, value) => Array.isArray(value) && rule.values.every((expected: unknown) => !value.includes(expected)),
-  },
-  arrayNotEmpty: {
-    defaultCode: 'ARRAY_NOT_EMPTY',
-    describe: (field) => `${field} must not be an empty array.`,
-    validate: (_rule, value) => Array.isArray(value) && value.length > 0,
-  },
-  arrayMinSize: {
-    defaultCode: 'ARRAY_MIN_SIZE',
-    describe: (field, rule) => `${field} must contain at least ${String(rule.value)} items.`,
-    validate: (rule, value) => Array.isArray(value) && value.length >= rule.value,
-  },
-  arrayMaxSize: {
-    defaultCode: 'ARRAY_MAX_SIZE',
-    describe: (field, rule) => `${field} must contain at most ${String(rule.value)} items.`,
-    validate: (rule, value) => Array.isArray(value) && value.length <= rule.value,
-  },
-  arrayUnique: {
-    defaultCode: 'ARRAY_UNIQUE',
-    describe: (field) => `${field} must contain unique values.`,
-    validate: (rule, value) => {
-      if (!Array.isArray(value)) return false;
-      const seen = new Set<unknown>();
-      for (const entry of value) {
-        const key = rule.selector ? rule.selector(entry) : entry;
-        if (seen.has(key)) return false;
-        seen.add(key);
-      }
-      return true;
-    },
-  },
-  custom: {
-    defaultCode: 'INVALID_FIELD',
-    describe: (field) => `${field} is invalid.`,
-    validate: () => true,
-  },
-};
 
 function createNestedDtoInstance<T>(target: Constructor<T>, rawValue: unknown, context?: NestedTraversalContext): T {
   if (rawValue instanceof target) {
@@ -479,12 +141,6 @@ function materializeNestedDtoValue<T>(target: Constructor<T>, rawValue: unknown,
   return createNestedDtoInstance(target, rawValue, context);
 }
 
-function getDtoBindingMap(target: Constructor): Map<MetadataPropertyKey, DtoFieldBindingMetadata> {
-  return new Map(
-    getDtoBindingSchema(target).map((entry: { propertyKey: MetadataPropertyKey; metadata: DtoFieldBindingMetadata }) => [entry.propertyKey, entry.metadata]),
-  );
-}
-
 function applyBindingValues(
   instance: Record<PropertyKey, unknown>,
   rawValue: Record<PropertyKey, unknown>,
@@ -525,60 +181,6 @@ function describeValidator(rule: DtoFieldValidationRule, field: string): { code:
     code: rule.code ?? (rule.kind === 'validatorjs' ? rule.validator.toUpperCase() : handler.defaultCode),
     message: rule.message ?? handler.describe(field, rule),
   };
-}
-
-function runValidatorJs(rule: Extract<DtoFieldValidationRule, { kind: 'validatorjs' }>, value: unknown): boolean {
-  if (rule.validator === 'latitude') {
-    return typeof value === 'number' && Number.isFinite(value) && value >= -90 && value <= 90;
-  }
-
-  if (rule.validator === 'longitude') {
-    return typeof value === 'number' && Number.isFinite(value) && value >= -180 && value <= 180;
-  }
-
-  if (typeof value !== 'string') {
-    return false;
-  }
-
-  switch (rule.validator) {
-    case 'alpha': return validator.isAlpha(value);
-    case 'alphanumeric': return validator.isAlphanumeric(value);
-    case 'ascii': return validator.isAscii(value);
-    case 'base64': return validator.isBase64(value);
-    case 'booleanString': return validator.isBoolean(value);
-    case 'currency': return validator.isCurrency(value, rule.args?.[0] as validator.IsCurrencyOptions | undefined);
-    case 'dataURI': return validator.isDataURI(value);
-    case 'dateString': return validator.isISO8601(value);
-    case 'decimal': return validator.isDecimal(value);
-    case 'email': return validator.isEmail(value, rule.args?.[0] as validator.IsEmailOptions | undefined);
-    case 'fqdn': return validator.isFQDN(value, rule.args?.[0] as validator.IsFQDNOptions | undefined);
-    case 'hexColor': return validator.isHexColor(value);
-    case 'hexadecimal': return validator.isHexadecimal(value);
-    case 'ip': return validator.isIP(value, rule.args?.[0] as '4' | '6' | undefined);
-    case 'isbn': return validator.isISBN(value, rule.args?.[0] as '10' | '13' | undefined);
-    case 'issn': return validator.isISSN(value);
-    case 'json': return validator.isJSON(value);
-    case 'jwt': return validator.isJWT(value);
-    case 'locale': return validator.isLocale(value);
-    case 'lowercase': return validator.isLowercase(value);
-    case 'magnetURI': return validator.isMagnetURI(value);
-    case 'matches': return validator.matches(value, rule.args?.[0] as string, rule.args?.[1] as string | undefined);
-    case 'mimeType': return validator.isMimeType(value);
-    case 'mobilePhone': return validator.isMobilePhone(value, rule.args?.[0] as validator.MobilePhoneLocale | validator.MobilePhoneLocale[] | undefined);
-    case 'mongoId': return validator.isMongoId(value);
-    case 'numberString': return validator.isNumeric(value);
-    case 'port': return validator.isPort(value);
-    case 'postalCode': return validator.isPostalCode(value, (rule.args?.[0] as validator.PostalCodeLocale | 'any' | undefined) ?? 'any');
-    case 'rgbColor': return validator.isRgbColor(value, rule.args?.[0] as boolean | undefined);
-    case 'rfc3339': return validator.isRFC3339(value);
-    case 'semVer': return validator.isSemVer(value);
-    case 'uppercase': return validator.isUppercase(value);
-    case 'url': return validator.isURL(value, rule.args?.[0] as validator.IsURLOptions | undefined);
-    case 'uuid': return validator.isUUID(value, rule.args?.[0] as validator.UUIDVersion | undefined);
-    case 'iso8601': return validator.isISO8601(value);
-    case 'latLong': return validator.isLatLong(value);
-    default: return false;
-  }
 }
 
 function buildIssue(fallback: { code: string; message: string }, field: string, source: ValidationIssue['source']): ValidationIssue {
